Move test to ItemAdditionalItemPropertyTest
